ENH: Add checks to prevent deletion of a specimen_id column

+++ qiita_db/metadata_template/base_metadata_template.py
@@ -647,6 +647,8 @@ class MetadataTemplate(qdb.base.QiitaObject):
             If the `column_name` doesn't exist
         QiitaDBOperationNotPermittedError
             If a the info file can't be updated
+            If the column_name is selected as a specimen_id_column in the
+            study.
         """
         if column_name not in self.categories():
             raise qdb.exceptions.QiitaDBColumnError(
@@ -654,6 +656,15 @@ class MetadataTemplate(qdb.base.QiitaObject):
         if not self.can_be_updated(columns={column_name}):
             raise qdb.exceptions.QiitaDBOperationNotPermittedError(
                 '%s cannot be deleted' % column_name)
+
+        # if a tube identifier column is selected disallow its deletion
+        specimen_id_column = qdb.study.Study(self.study_id).specimen_id_column
+        if specimen_id_column == column_name:
+            raise qdb.exceptions.QiitaDBOperationNotPermittedError(
+                    '"%s" cannot be deleted, this column is currently selected'
+                    ' as the tube identifier (specimen_id_column)' %
+                    column_name)
+
         with qdb.sql_connection.TRN:
             sql = 'ALTER TABLE qiita.%s%d DROP COLUMN %s' % (
                 self._table_prefix, self._id, column_name)
